Controllers, Closes #17

+++ backend/src/main/java/com/rentacar/backend/services/VehiculoService.java
@@ -2,19 +2,27 @@ package com.rentacar.backend.services;
 
 import com.rentacar.backend.entities.SucursalEntity;
 import com.rentacar.backend.entities.VehiculoEntity;
+import com.rentacar.backend.repositories.ReservaRepository;
 import com.rentacar.backend.repositories.VehiculoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.regex.Pattern;
 
 @Service
 public class VehiculoService {
+    private final VehiculoRepository vehiculoRepository;
+    private final ReservaRepository reservaRepository;
+
     @Autowired
-    private VehiculoRepository vehiculoRepository;
+    public VehiculoService(VehiculoRepository vehiculoRepository, ReservaRepository reservaRepository) {
+        this.vehiculoRepository = vehiculoRepository;
+        this.reservaRepository = reservaRepository;
+    }
 
     /**
      * Permite crear un nuevo vehículo y almacenarlo en la base de datos.
@@ -25,7 +33,7 @@ public class VehiculoService {
      * @param precio precio de arriendo
      * @return Entidad Vehículo creada
      */
-    VehiculoEntity crearVehiculo(String marca, String modelo, String acriss,
+    public VehiculoEntity crearVehiculo(String marca, String modelo, String acriss,
                                  String patente, BigDecimal precio) {
         // Sanitización de datos
 
@@ -64,13 +72,26 @@ public class VehiculoService {
         return patternAcriss.matcher(acriss.toUpperCase()).matches();
     }
 
+    /**
+     * Permite obtener todos los vehículos en la base de datos.
+     * @return Lista de todos los vehículos
+     */
+    public List<VehiculoEntity> obtenerVehiculos() {
+        return vehiculoRepository.findAll();
+    }
+
+    public VehiculoEntity obtenerVehiculoPorId(Long id) {
+        Optional<VehiculoEntity> vehiculo = vehiculoRepository.findById(id);
+        if (vehiculo.isEmpty()) throw new NoSuchElementException("Vehiculo no encontrado");
+        return vehiculo.get();
+    }
 
     /**
      * Permite obtener un vehículo con una patente determinada.
      * @param patente Patente del vehículo a buscar
      * @return Vehículo con la patente buscada, si existe.
      */
-    Optional<VehiculoEntity> obtenerVehiculoPorPatente(String patente) {
+    public Optional<VehiculoEntity> obtenerVehiculoPorPatente(String patente) {
         return vehiculoRepository.findByPatente(patente);
     }
 
@@ -79,7 +100,7 @@ public class VehiculoService {
      * @param marca Marca de los vehículos
      * @return Lista de vehículos de esa marca
      */
-    List<VehiculoEntity> obtenerVehiculosPorMarca(String marca) {
+    public List<VehiculoEntity> obtenerVehiculosPorMarca(String marca) {
         return vehiculoRepository.findByMarca(marca);
     }
 
@@ -90,7 +111,7 @@ public class VehiculoService {
      * @param acriss Código ACRISS
      * @return Lista de vehículos con código ACRISS acorde a lo pedido
      */
-    List<VehiculoEntity> obtenerVehiculosPorAcriss(String acriss) {
+    public List<VehiculoEntity> obtenerVehiculosPorAcriss(String acriss) {
         if (!validarAcriss(acriss))
             throw new IllegalArgumentException("Código ACRISS no válido");
         return vehiculoRepository.findByAcrissLike(acriss.toUpperCase());
@@ -101,8 +122,10 @@ public class VehiculoService {
      * @param sucursal Sucursal a buscar
      * @return Lista de vehículos en esa sucursal
      */
-    List<VehiculoEntity> obtenerVehiculosPorSucursal(SucursalEntity sucursal) {
-        return vehiculoRepository.findBySucursal(sucursal);
+    public List<VehiculoEntity> obtenerVehiculosPorSucursal(SucursalEntity sucursal) {
+        List<VehiculoEntity> v = vehiculoRepository.findBySucursal(sucursal);
+        if (v.isEmpty()) throw new RuntimeException("No hay vehiculos en la sucursal " + sucursal.getNombre());
+        return v;
     }
 
     /**
@@ -111,7 +134,7 @@ public class VehiculoService {
      * @param max Precio máximo de arriendo
      * @return Lista de vehículos con precio de arriendo entre min y max
      */
-    List<VehiculoEntity> obtenerVehiculosConPrecioEntre(BigDecimal min, BigDecimal max) {
+    public List<VehiculoEntity> obtenerVehiculosConPrecioEntre(BigDecimal min, BigDecimal max) {
         if (min.compareTo(BigDecimal.ZERO) < 0 || max.compareTo(BigDecimal.ZERO) < 0)
             throw new IllegalArgumentException("Precio inválido");
 
@@ -123,8 +146,10 @@ public class VehiculoService {
      * @param estado Estado de los vehículos
      * @return Lista de vehículos en ese estado
      */
-    List<VehiculoEntity> obtenerVehiculosPorEstado(String estado) {
-        return vehiculoRepository.findByEstado(estado);
+    public List<VehiculoEntity> obtenerVehiculosPorEstado(String estado) {
+        List<VehiculoEntity> v = vehiculoRepository.findByEstado(estado);
+        if (v.isEmpty()) throw new RuntimeException("No se encontraron vehiculos en ese estado");
+        return v;
     }
 
     /**
@@ -133,9 +158,21 @@ public class VehiculoService {
      * @param estado Estado nuevo del vehículo
      * @return Vehículo actualizado con ese nuevo estado
      */
-    VehiculoEntity actualizarEstadoVehiculoPorId(Long vehiculoId, String estado) {
+    public VehiculoEntity actualizarEstadoVehiculoPorId(Long vehiculoId, String estado) {
         VehiculoEntity vehiculo = vehiculoRepository.findById(vehiculoId).orElseThrow();
         vehiculo.setEstado(estado);
+        return vehiculoRepository.save(vehiculo);
+    }
+
+    /**
+     * Permite actualizar el precio de arriendo de un vehiculo dado su ID
+     * @param vehiculoId ID del vehiculo a actualizar
+     * @param precio Precio nuevo del vehiculo
+     * @return Vehiculo actualizado con ese nuevo precio
+     */
+    public VehiculoEntity actualizarPrecioArriendoVehiculoPorId(Long vehiculoId, BigDecimal precio) {
+        VehiculoEntity vehiculo = vehiculoRepository.findById(vehiculoId).orElseThrow();
+        vehiculo.setPrecioArriendo(precio);
         return vehiculoRepository.save(vehiculo);
     }
 
@@ -143,8 +180,11 @@ public class VehiculoService {
      * Permite eliminar un vehículo de la base de datos dado su ID
      * @param vehiculoId ID del vehículo a eliminar
      */
-    void eliminarVehiculoPorId(Long vehiculoId) {
-        vehiculoRepository.findById(vehiculoId).orElseThrow();
-        vehiculoRepository.deleteById(vehiculoId);
+    public void eliminarVehiculoPorId(Long vehiculoId) {
+        VehiculoEntity vehiculo = vehiculoRepository.findById(vehiculoId).orElseThrow();
+        if (!reservaRepository.findByVehiculo(vehiculo).isEmpty())
+            throw new RuntimeException("El vehiculo" + vehiculoId + "tiene reservas pendientes");
+        else
+            vehiculoRepository.deleteById(vehiculoId);
     }
 }
